fix: Retrieve chart version from current release branch

+++ Makefile
@@ -42,12 +42,14 @@ update_dependencies: setup_helm_repos helm_dep_up update_internals_versions upda
 
 .PHONY: get_release_notes
 get_release_notes:
+	# Obtain current and previous chart versions
+	$(eval VERSION_NEW=$(shell git branch --remote --sort version:refname | grep -Eo "release-[0-9]+\.[0-9]+\.[0-9]+$$" | tail -n 1 | cut -d - -f 2))
+	$(eval VERSION_OLD=$(shell git branch --remote --sort version:refname | grep -Eo "release-[0-9]+\.[0-9]+\.[0-9]+$$" | tail -n 2 | head -n 1 | cut -d - -f 2))
+	if [ -z "${VERSION_NEW}" ] || [ -z "${VERSION_OLD}" ]; then echo "Can't find the current or previous chart version."; exit 1; fi
 	# Clone codacy/codacy-tools-release-notes
 	if [ -d codacy-tools-release-notes ]; then rm -rf codacy-tools-release-notes; fi
 	git clone -b clean/refactor-scripts-DOCS-206 git@github.com:codacy/codacy-tools-release-notes.git
 	# Fetch updated codacy/chart tags
 	git fetch --all --tags --force
 	# Generate changelog and release notes
-	codacy-tools-release-notes/run.sh \
-		$(shell git tag --sort v:refname | grep -E "^[0-9]+\.[0-9]+\.[0-9]+$$" | tail -n 1) \
-		$(shell git tag --sort v:refname | grep -E "^[0-9]+\.[0-9]+\.[0-9]+$$" | tail -n 2 | head -n 1)
+	codacy-tools-release-notes/run.sh ${VERSION_NEW} ${VERSION_OLD}
